requirements.txt
aenum==3.1.11
affine==2.3.1
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==7.1.2
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.5
contourpy==1.0.6
coreapi==2.3.3
coreschema==0.0.4
cors==1.0.1
coverage==6.5.0
cycler==0.11.0
dataclasses-json==0.5.7
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
django-jazzmin==2.6.0
django-leaflet==0.28.3
django-modeltranslation==0.18.9
django-schema-graph==2.2.1
django-simple-history==3.2.0
djangorestframework==3.12.4
djangorestframework-gis==1.0
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==15.1.1
filelock==3.8.0
Fiona==1.8.22
fonttools==4.38.0
frozenlist==1.3.1
future==0.18.2
GDAL==3.2.2
geojson==3.0.0
geomet==1.0.0
geopandas==0.12.2
geoserver-rest==2.4.1
gevent==22.8.0
greenlet==1.1.3.post0
html2text==2020.1.16
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
landsatxplore==0.13.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.6.2
multidict==6.0.2
munch==2.5.0
mypy-extensions==1.0.0
numpy==1.24.0
oauthlib==3.2.2
OWSLib==0.27.2
packaging==21.3
pandas==1.5.2
Pillow==9.2.0
psycopg2-binary==2.9.5
Pygments==2.14.0
pyparsing==3.0.9
pyproj==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.4
python-weather==0.4.2
pytz==2022.4
PyYAML==6.0
rasterio==1.3.4
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
seaborn==0.12.1
sentinelhub==3.8.3
sentinelsat==1.1.1
serializers==0.2.4
Shapely==1.8.5.post1
six==1.16.0
snuggs==1.4.7
sqlparse==0.4.3
tifffile==2023.1.23.1
tldextract==3.4.0
tqdm==4.64.1
typing-inspect==0.8.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.12
utm==0.7.0
xmltodict==0.13.0
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.5.0
